check that player is set before using it
this check is not necessary with EntityPlayerMP but from FakePlayer it
is possible that the field has not been set. In OpenComputesr, our robot
had not set this field (we are now) but I thought it would be nice to
have this fix on both sides. This NRE was causing our robot to fail to
pickup items through a pickup event listener coming back to this code.

We could have instead added an instanceOf check on the player (as null
is not an instance of anything in java), but, I thought it would be
nice for FakePlayers to be able to mimic when possible.